LayoutTests/imported/w3c/web-platform-tests/mediacapture-fromelement/creation-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]
CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]
CONSOLE MESSAGE: Unhandled Promise Rejection: [object DOMError]

Harness Error (FAIL), message = The operation is not supported.

PASS video-only
FAIL Untitled assert_true: expected true got false
PASS audio-only
FAIL Untitled 1 assert_true: expected true got false
PASS video+audio
FAIL Untitled 2 assert_true: expected true got false
